+++ library/core/src/main/java/com/google/android/exoplayer2/drm/DefaultDrmSession.java
@@ -257,6 +257,11 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
   }
 
   @Override
+  public final UUID getSchemeUuid() {
+    return uuid;
+  }
+
+  @Override
   public final @Nullable ExoMediaCrypto getMediaCrypto() {
     return mediaCrypto;
   }